// dojo/src/components/constants/types.ts
export interface FormData {
  traineeInfo: {
    name: string;
    id: string;
    empNo: string;
    stationName: string;
    lineName: string;
    processName: string;
    revisionDate: string;
    doi: string;
    trainerName: string;
    trainerLine: string;
  };
  dailyScores: {
    [topicId: string | number]: {
      [dayId: string | number]: string;
    };
  };
  signatures: {
    preparedBy: string;
    approvedBy: string;
    engineerJudge: string;
  };
}